backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
PyMuPDF
python-multipart
